Move network topology tests into own file
